fix a bug that caused players list to be empty
Making the view models testable required the dispatcher to be
injected. Previously, `.launchIn` was used to launch the job.

`.launchIn` does not provide the option to provide a context,
therefore a switch to `.launch(dispatcher)` was needed. During
this change, I made a mistake and did not `.collect()` the Flow
emitted after launching the job. This caused the Flow to never be
collected.